[Fix/#483] 기억빛 gradientView 컴포넌트로 분리
